Adding comments to propTypes; Moving week logic to Week component; Removing WeekCalendar component
